pyproject.toml
[tool.poetry]
name = "langchain-tutorials"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.7"
python-dotenv = "^1.0.1"
ruff = "^0.7.3"
langchain = "^0.3.7"
langchain-openai = "^0.2.6"

[tool.poetry.group.simple_llm.dependencies]
langserve = {extras = ["all"], version = "^0.3.0"}

[tool.poetry.group.chatbot.dependencies]
langchain-core = "^0.3.15"
langgraph = "^0.2.45"

[tool.poetry.group.vector_stores.dependencies]
langchain-chroma = "^0.1.4"

[tool.poetry.group.agent.dependencies]
langchain-community = "^0.3.5"
langgraph = "^0.2.45"
langchain-anthropic = "^0.2.4"
tavily-python = "^0.5.0"
langgraph-checkpoint-sqlite = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
